feat: add nice control window
